ZENYX/requirements.txt
# requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
redis==4.5.5
aioredis==2.0.1
requests==2.28.2
pydantic==1.10.8
python-dateutil==2.8.2
matplotlib==3.7.1
pillow==9.5.0